nlp-service/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
spacy>=3.7.0
PyPDF2>=3.0.1
python-docx>=1.1.0
pymongo>=4.6.1
python-multipart>=0.0.6
pydantic>=2.5.0
scikit-learn>=1.4.0
numpy>=1.26.0
